Car_Repair/django/requirements.txt
asgiref==3.6.0
bcrypt==4.0.1
bootstrap4==0.1.0
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
Django==4.1.4
django-phonenumber-field==7.0.1
django-phonenumbers==1.0.1
dnspython==2.3.0
efficientnet-pytorch==0.7.1
elastic-transport==8.4.0
elasticsearch==8.6.0
filelock==3.9.0
fonttools==4.38.0
huggingface-hub==0.11.1
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.3
munch==2.5.0
mysqlclient==2.1.1
numpy==1.24.1
opencv-python==4.7.0.68
packaging==23.0
pandas==1.5.2
phonenumbers==8.13.3
phonenumberslite==8.13.3
Pillow==9.4.0
pretrainedmodels==0.7.4
pymongo==4.3.3
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
scikit-learn==1.2.0
scipy==1.10.0
segmentation-models-pytorch==0.3.2
six==1.16.0
sqlparse==0.4.3
threadpoolctl==3.1.0
timm==0.6.12
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
xgboost==1.7.3
